Add FPU arithmetic, interrupt dispatch, and exception handling
Complete COP1 implementation: single/double-precision arithmetic (ADD,
SUB, MUL, DIV, SQRT, ABS, MOV, NEG), rounding/truncation to integer
(ROUND, TRUNC, CEIL, FLOOR for W and L formats), format conversions
(CVT.S/D/W/L), and FP compare instructions (C.cond with proper
unordered/NaN handling). Also adds MFC1/MTC1/CFC1/CTC1, BC1T/BC1F,
and LWC1/SWC1/LDC1/SDC1.

Implements interrupt dispatch in the CPU step loop: IP2 now reflects
MI state each cycle, and when interrupt conditions are met (IE=1,
EXL=0, ERL=0, unmasked IP bits), the CPU vectors to the exception
handler at 0x80000180 with proper EPC/Cause/Status updates.

Zelda OoT now boots past initial code and hits TLB-mapped addresses.

+++ crates/n64-core/src/cpu/instruction.rs
@@ -22,7 +22,7 @@ impl Instruction {
     pub fn rs(self) -> usize { ((self.0 >> 21) & 0x1F) as usize }
     pub fn rt(self) -> usize { ((self.0 >> 16) & 0x1F) as usize }
     pub fn rd(self) -> usize { ((self.0 >> 11) & 0x1F) as usize }
-    pub fn sa(self) -> u32 { (self.0 >> 6) & 0x1F }
+    pub fn sa(self) -> usize { ((self.0 >> 6) & 0x1F) as usize }
     pub fn funct(self) -> u32 { self.0 & 0x3F }
     pub fn imm(self) -> u16 { self.0 as u16 }
     pub fn imm_sign_ext(self) -> u64 { self.imm() as i16 as i64 as u64 }
